Correct __readmsr argument to be unsigned long.
git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@261 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-crt/misc/io_intrin.c
@@ -277,7 +277,7 @@ void __outdwordstring(unsigned short Port,unsigned long *Buffer,unsigned long Co
  
   #endif
 
-  unsigned __int64 __readmsr(int msr)
+  unsigned __int64 __readmsr(unsigned long msr)
   {
       unsigned __int64 val1, val2;
        __asm__ __volatile__(
